LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/upgrade-insecure-css-in-iframe.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function (e) {
    window.data = e.data;
    shouldBe("window.data", "'done'");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);
</script>
<p>This test loads a secure iframe that loads an insecure style sheet.
We should upgrade the CSS request to HTTPS, and thereby avoid triggering
a mixed content callback.</p>
<iframe src="https://127.0.0.1:8443/security/contentSecurityPolicy/upgrade-insecure-requests/resources/frame-with-insecure-css.html"></iframe>
</body>
</html>
